Move PixelMaze to lab
