+++ lab/PixelMaze/Sprite.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { Point } from "../pga2d/objects.js";
+import { Point } from "../../pga2d/objects.js";
 import { ImageFrames } from "./ImageFrames.js";
 import { Sprite } from "./Sprite.js";
