+++ lab/PixelMaze/index.html
@@ -5,14 +5,14 @@
     <script
       language="javascript"
       type="text/javascript"
-      src="../libraries/p5.min.js"
+      src="../../libraries/p5.min.js"
     ></script>
     <script type="module">
       import { sketch } from "./PixelMaze.js";
-      import "../sketchlib/sketchbook-navbar.js";
+      import "../../sketchlib/sketchbook-navbar.js";
       const p5_sketch = new p5(sketch, "sketch");
     </script>
-    <link rel="stylesheet" href="../styles/layout.css" />
+    <link rel="stylesheet" href="../../styles/layout.css" />
   </head>
 
   <body>
