+++ lab/PixelMaze/sprites/placeholder-character.png
