+++ lab/index.html
@@ -51,6 +51,7 @@
         <li><a href="./DoubleSpring/">Double Spring</a></li>
         <li><a href="./PlanarCurves/">Planar Curves</a></li>
         <li><a href="./RasterTangles/">Raster Tangles</a></li>
+        <li><a href="./PixelMaze/">Pixel Maze</a></li>
       </ul>
     </div>
   </body>
