+++ sketchlib/links.js
@@ -1,10 +1,6 @@
 // Links for all of the sketches. Links are added to the page in this order
 export const LINKS = [
   {
-    title: "Pixel Maze",
-    year: "2025",
-  },
-  {
     title: "Anabaena",
     year: "2025",
   },
